Time slot cook #15 (#33)
* Preparation TimeSlots #15

* Fonctions TimeSlots #15

* Exception + Comparaison TimeSlots #15

* Tests TimeSlots #15

* Merge Time Slots

+++ src/main/java/fr/unice/polytech/recipe/Cookie.java
@@ -18,7 +18,7 @@ public class Cookie {
     private Double price;
     @Getter
     @Setter
-    private Double cookingTime;
+    private int cookingTime;
     @Getter
     @Setter
     private Cooking cooking;
